+++ src/main/java/com/soolsul/soolsulserver/post/business/PostQueryService.java
@@ -7,10 +7,10 @@ import com.soolsul.soolsulserver.bar.exception.BarNotFoundException;
 import com.soolsul.soolsulserver.bar.persistence.BarQueryRepository;
 import com.soolsul.soolsulserver.bar.presentation.dto.BarLookupResponse;
 import com.soolsul.soolsulserver.location.response.LocationSquareRangeCondition;
-import com.soolsul.soolsulserver.post.business.dto.PostDetailLikeResponse;
-import com.soolsul.soolsulserver.post.business.dto.PostDetailStoreResponse;
-import com.soolsul.soolsulserver.post.business.dto.PostDetailUserResponse;
-import com.soolsul.soolsulserver.post.business.dto.PostLookupRequest;
+import com.soolsul.soolsulserver.post.business.dto.response.PostDetailLikeResponse;
+import com.soolsul.soolsulserver.post.business.dto.response.PostDetailStoreResponse;
+import com.soolsul.soolsulserver.post.business.dto.response.PostDetailUserResponse;
+import com.soolsul.soolsulserver.post.business.dto.request.PostLookupRequest;
 import com.soolsul.soolsulserver.post.domain.Post;
 import com.soolsul.soolsulserver.post.domain.PostPhoto;
 import com.soolsul.soolsulserver.post.domain.PostRepository;
