+++ src/main/java/com/soolsul/soolsulserver/post/business/dto/request/PostLookupRequest.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.post.business.dto;
+package com.soolsul.soolsulserver.post.business.dto.request;
 
 import com.soolsul.soolsulserver.user.auth.repository.dto.UserLookUpResponse;
 import com.soolsul.soolsulserver.post.domain.Post;
