+++ src/main/java/com/soolsul/soolsulserver/post/business/dto/response/PostDetailStoreResponse.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.post.business.dto;
+package com.soolsul.soolsulserver.post.business.dto.response;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
