+++ src/main/java/com/soolsul/soolsulserver/post/presentation/dto/PostDetailResponse.java
@@ -2,9 +2,9 @@ package com.soolsul.soolsulserver.post.presentation.dto;
 
 import com.soolsul.soolsulserver.user.auth.repository.dto.UserLookUpResponse;
 import com.soolsul.soolsulserver.bar.presentation.dto.BarLookupResponse;
-import com.soolsul.soolsulserver.post.business.dto.PostDetailLikeResponse;
-import com.soolsul.soolsulserver.post.business.dto.PostDetailStoreResponse;
-import com.soolsul.soolsulserver.post.business.dto.PostDetailUserResponse;
+import com.soolsul.soolsulserver.post.business.dto.response.PostDetailLikeResponse;
+import com.soolsul.soolsulserver.post.business.dto.response.PostDetailStoreResponse;
+import com.soolsul.soolsulserver.post.business.dto.response.PostDetailUserResponse;
 import com.soolsul.soolsulserver.post.domain.Post;
 import org.hibernate.validator.constraints.URL;
