+++ src/test/java/com/soolsul/soolsulserver/unit/post/business/PostQueryFacadeSpecTest.java
@@ -5,7 +5,7 @@ import com.soolsul.soolsulserver.user.auth.repository.dto.UserLookUpResponse;
 import com.soolsul.soolsulserver.bar.persistence.BarQueryRepository;
 import com.soolsul.soolsulserver.bar.presentation.dto.BarLookupResponse;
 import com.soolsul.soolsulserver.post.business.PostQueryService;
-import com.soolsul.soolsulserver.post.business.dto.PostLookupRequest;
+import com.soolsul.soolsulserver.post.business.dto.request.PostLookupRequest;
 import com.soolsul.soolsulserver.post.domain.Post;
 import com.soolsul.soolsulserver.post.domain.PostPhoto;
 import com.soolsul.soolsulserver.post.presentation.dto.PostDetailResponse;
